goprotobuf: Make edge-case int32 encoding more compact, and the same as C++.
Fixes #39.

R=r
CC=golang-dev
https://codereview.appspot.com/10461043

+++ proto/encode.go
@@ -235,7 +235,7 @@ func (o *Buffer) enc_int32(p *Properties, base structPointer) error {
 	}
 	x := word32_Get(v)
 	o.buf = append(o.buf, p.tagcode...)
-	p.valEnc(o, uint64(int32(x)))
+	p.valEnc(o, uint64(x))
 	return nil
 }
